Eliminate SQL

+++ src/main/java/proteomics/Math/CalScore.java
@@ -3,22 +3,22 @@ package proteomics.Math;
 
 public class CalScore {
 
-	private double[][] alignment_matrix = null;
+	private float[][] alignment_matrix = null;
 
-	public CalScore(double[][] exp_matrix, double[] theo_vector, double xcorr_tolerance) {
-		alignment_matrix = new double[2][exp_matrix[0].length + theo_vector.length];
+	public CalScore(float[][] exp_matrix, float[] theo_vector, float xcorr_tolerance) {
+		alignment_matrix = new float[2][exp_matrix[0].length + theo_vector.length];
 
 		int start = 0;
 		int idx = 0;
-		for (double theo_mz : theo_vector) {
+		for (float theo_mz : theo_vector) {
 			if (exp_matrix[0][0] - theo_mz > xcorr_tolerance) {
 				continue;
 			}
 			
 			int j = start;
 			while (j < exp_matrix[0].length) {
-				double exp_mz = exp_matrix[0][j];
-				double temp = exp_mz - theo_mz;
+				float exp_mz = exp_matrix[0][j];
+				float temp = exp_mz - theo_mz;
 				if ((temp >= -xcorr_tolerance) && (temp <= xcorr_tolerance)) {
 					alignment_matrix[0][idx] = exp_mz;
 					alignment_matrix[1][idx] = exp_matrix[1][j];
@@ -32,8 +32,8 @@ public class CalScore {
 		}
 	}
 
-	public double cal_dot_product() {
-		double dot_value = 0;
+	public float cal_dot_product() {
+		float dot_value = 0;
 		for (int i = 0; i < alignment_matrix[0].length; ++i) {
 			dot_value += alignment_matrix[1][i];
 		}
